fix:minor fixes

+++ lib/features/task_entity_service/data/models/task_entity_service.dart
@@ -198,11 +198,11 @@ class Rating with _$Rating {
 class RatingStats with _$RatingStats {
   const factory RatingStats({
     @JsonKey(name: "average_rating") double? averageRating,
-    double? five,
-    double? four,
-    double? three,
-    double? two,
-    double? one,
+    int? five,
+    int? four,
+    int? three,
+    int? two,
+    int? one,
     @JsonKey(name: "total_counts") int? totalCounts,
   }) = _RatingStats;
